@@ -1,47 +1,21 @@
 package com.in28minutes.junit;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.*;
 
-import org.junit.After;
-import org.junit.AfterClass;
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
-public class MyMathTest {
-	MyMath myMath = new MyMath();
+class MyMathTest {
 
-	@Before
-	public void before() {
-		System.out.println("Before");
-	}
-
-	@After
-	public void after() {
-		System.out.println("After");
-	}
-
-	@BeforeClass
-	public static void beforeClass() {
-		System.out.println("Before Class");
-	}
-
-	@AfterClass
-	public static void afterClass() {
-		System.out.println("After Class");
-	}
-
-	// MyMath.sum
-	// 1,2,3 => 6
+	private MyMath math = new MyMath();
+	
 	@Test
-	public void sum_with3numbers() {
-		System.out.println("Test1");
-		assertEquals(6, myMath.sum(new int[] { 1, 2, 3 }));
+	void calculateSum_ThreeMemberArray() {		
+		assertEquals(6, math.calculateSum(new int[] {1,2,3}));
 	}
 
 	@Test
-	public void sum_with1number() {
-		System.out.println("Test2");
-		assertEquals(3, myMath.sum(new int[] { 3 }));
+	void calculateSum_ZeroLengthArray() {		
+		assertEquals(0, math.calculateSum(new int[] {}));
 	}
+
 }
