@@ -1,24 +1,9 @@
-package com.in28minutes.learnspringboot.courses.bean;
+package com.in28minutes.springboot.learnspringboot;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-
-@Entity
 public class Course {
-
-	@Id
-	@GeneratedValue
 	private long id;
-	
-	//@Column(name="course_name")
 	private String name;
-
 	private String author;
-	
-	public Course() {
-	}
 
 	public Course(long id, String name, String author) {
 		super();
