@@ -1,4 +1,4 @@
-package com.in28minutes.learnspringboot;
+package com.in28minutes.springboot.learnspringboot;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
